% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpEC_trait1_hidden_essaie.R
\name{DAISIE_DE_logpEC_trait1_hidden}
\alias{DAISIE_DE_logpEC_trait1_hidden}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_logpEC_trait1_hidden(
  brts,
  missnumspec,
  parameter,
  phy,
  traits,
  num_hidden_traits,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  see_ancestral_states = TRUE,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45",
  rhs_func = loglik_hidden_rhs
)
}
\arguments{
\item{brts}{branching times}

\item{missnumspec}{number of missing species}

\item{parameter}{parameters}

\item{phy}{phy}

\item{traits}{traits}

\item{cond}{conditioning, default = "proper_cond"}

\item{root_state_weight}{root weight, default = "proper_weights"}

\item{see_ancestral_states}{recover the ancestral states}

\item{atol}{absolute tolerance}

\item{rtol}{relative tolerance}

\item{methode}{method of integration}

\item{setting_calculation}{used in ML}
}
\description{
This function calculates something we can verify with DAISIE
}
