% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpEC_trait1_hidden.R
\name{DAISIE_DE_logpEC_trait1_hidden_old}
\alias{DAISIE_DE_logpEC_trait1_hidden_old}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_logpEC_trait1_hidden_old(
  brts,
  missnumspec,
  parameter,
  phy,
  traits,
  sampling_fraction,
  num_hidden_traits,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  see_ancestral_states = TRUE,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45",
  rhs_func = loglik_hidden_rhs
)
}
\arguments{
\item{brts}{branching times, being: island age, colonization time, branching times}

\item{missnumspec}{missing number of species}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{phy}{phylogenetic tree of class \code{phylo}, rooted and with
branch lengths. Alternatively, multiple phylogenetic trees can be provided
as the \code{multiPhylo} class.}

\item{traits}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
\code{vignette("starting_secsse", package = "secsse")}. When providing a
\code{multiPhylo} set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a \code{multiPhylo} object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{num_hidden_traits}{number of hidden traits}

\item{cond}{condition on the existence of a node root: \code{"maddison_cond"},
\code{"proper_cond"} (default). For details, see vignette.}

\item{root_state_weight}{the method to weigh the states:
\code{"maddison_weights"}, \code{"proper_weights"} (default) or \code{"equal_weights"}.
It can also be specified for the root state: the vector \code{c(1, 0, 0)}
indicates state 1 was the root state. When
using a \code{multiPhylo} object, root_state_weight should be list where each
entry in the list corresponds to the root_state_weight for each tree.}

\item{see_ancestral_states}{Boolean for whether the ancestral states should
be shown? Defaults to \code{FALSE}.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{rhs_func}{rhs function used in R implementation}
}
\description{
This function calculates something we can verify with DAISIE, using hidden traits
}
