% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_loglik_CS.R
\name{DAISIE_DE_trait_loglik_CS}
\alias{DAISIE_DE_trait_loglik_CS}
\title{apply loglikelihood to a dataset}
\usage{
DAISIE_DE_trait_loglik_CS(
  parameter,
  datalist,
  methode = "lsodes",
  rcpp_methode = "odeint::bulirsch_stoer",
  atol = 1e-15,
  rtol = 1e-15,
  num_observed_states,
  num_hidden_states,
  cond = 1,
  num_threads = 1,
  verbose = FALSE,
  use_Rcpp = 0
)
}
\arguments{
\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{datalist}{a list containing in the first entry: island age and the
number of species not present, then from entry 2 until n it contains
individual datasets for which to estimate the loglikelihood together, each
dataset has the following fields: 1) colonist_name, 2) b ranching_times,
3) stac, 4) missing_species, 5) type1or2, 6) traits, 7) sampling fraction,
8) phylogeny}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
\code{"odeint::runge_kutta_cash_karp54"}, \code{"odeint::runge_kutta_fehlberg78"},
\code{"odeint::runge_kutta_dopri5"}, \code{"odeint::bulirsch_stoer"} and
\code{"odeint::runge_kutta4"}. Default method is: \code{"odeint::bulirsch_stoer"}.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{cond}{condition on the existence of a node root: \code{"maddison_cond"},
\code{"proper_cond"} (default). For details, see vignette.}

\item{num_threads}{number of threads to be used. Default is one thread.}

\item{verbose}{sets verbose output; default is \code{TRUE} when \code{optimmethod} is
\code{"simplex"}. If \code{optimmethod} is set to \code{"simplex"}, then even if set to
\code{FALSE}, optimizer output will be shown.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
umbrella function
}
