% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logp0.R
\name{DAISIE_DE_trait_logp0}
\alias{DAISIE_DE_trait_logp0}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logp0(
  brts,
  parameter,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45"
)
}
\arguments{
\item{atol}{absolute tolerance}

\item{rtol}{relative tolerance}

\item{methode}{method of integration}
}
\description{
This function compute the likelihood that all species that colonize the island
have gone extinct prior to the present.
}
\examples{
#Load DAISIE package and data
library(DAISIE)
data("Galapagos_datalist")

parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0
)

parameter <- list(2.546591, 2.678781, 0.009326754, 1.008583, matrix(c(0), nrow = 1), 0 )

# Compute log‐likelihood under the DE‐trait model
DAISIE_DE_trait_logp0(
  brts            = datalist[[1]]$island_age,
  parameter           = parameter,
  atol                = 1e-10,
  rtol                = 1e-10,
  methode             = "ode45"
)
}
