% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpEC_hidden.R
\name{DAISIE_DE_trait_logpEC_hidden}
\alias{DAISIE_DE_trait_logpEC_hidden}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpEC_hidden(
  brts,
  parameter,
  phy,
  traits,
  sampling_fraction,
  trait_mainland_ancestor = FALSE,
  num_observed_states,
  num_hidden_states,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  see_ancestral_states = TRUE,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45",
  rhs_func = loglik_hidden_rhs
)
}
\arguments{
\item{brts}{branching times}

\item{parameter}{parameters}

\item{phy}{phylogeny}

\item{traits}{traits}

\item{trait_mainland_ancestor}{trait of the species at the stem}

\item{num_observed_states}{number of observed traits}

\item{num_hidden_states}{number of hidden traits}

\item{cond}{conditioning, default = "proper_cond"}

\item{root_state_weight}{root weight, default = "proper_weights"}

\item{see_ancestral_states}{recover the ancestral states}

\item{atol}{absolute tolerance}

\item{rtol}{relative tolerance}

\item{methode}{method of integration}

\item{missnumspec}{number of missing species}

\item{setting_calculation}{used in ML}
}
\description{
this function calculate the likelihood of observing a clade with specified species trait states,
and for which the estimated colonization is known.
}
\examples{
library(secsse)
library(DAISIE)
data("NewZealand_birds_datalist")
datalist <- NewZealand_birds_datalist
i <- 28
phy <- DDD::brts2phylo(datalist[[i]]$branching_times[-c(1, 2)])
traits <- sample(c(0,0), length(phy$tip.label), replace = TRUE)
sampling_fraction <- sample(c(1,1), length(phy$tip.label), replace = TRUE)

parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0
)

parameter <- list(0.1255591, 0.1910084, 0.004697986, 0.326698, matrix(c(0), nrow = 1), 0 )

DAISIE_DE_trait_logpEC_hidden(
  brts                  = datalist[[i]]$branching_times,
  phy                   = phy,
  traits                = traits,
  sampling_fraction     = sampling_fraction,
  trait_mainland_ancestor = FALSE,
  parameter             = parameter,
  num_observed_states   = 2,
  num_hidden_states     = 2,
  cond                  = "proper_cond",
  root_state_weight     = "proper_weights",
  see_ancestral_states  = TRUE,
  atol                  = 1e-10,
  rtol                  = 1e-10,
  methode               = "ode45",
  rhs_func              = loglik_hidden_rhs
)
}
