% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpEC_mainland_hidden.R
\name{DAISIE_DE_trait_logpEC_mainland_hidden}
\alias{DAISIE_DE_trait_logpEC_mainland_hidden}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpEC_mainland_hidden(
  brts,
  missnumspec,
  parameter,
  phy,
  traits,
  trait_mainland_ancestor,
  sampling_fraction,
  num_observed_states,
  num_hidden_states,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  see_ancestral_states = TRUE,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45",
  rhs_func = loglik_hidden_rhs_mainland()
)
}
\arguments{
\item{brts}{branching times}

\item{missnumspec}{number of missing species}

\item{parameter}{parameters}

\item{phy}{phy}

\item{traits}{traits}

\item{trait_mainland_ancestor}{trait of the species at the stem}

\item{num_observed_states}{number of observed traits}

\item{num_hidden_states}{number of hidden traits}

\item{cond}{conditioning, default = "proper_cond"}

\item{root_state_weight}{root weight, default = "proper_weights"}

\item{see_ancestral_states}{recover the ancestral states}

\item{atol}{absolute tolerance}

\item{rtol}{relative tolerance}

\item{methode}{method of integration}

\item{setting_calculation}{used in ML}
}
\description{
this function calculates the likelihood of observing a clade on an island
that coexists with its mainland ancestors.
}
