% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpEC_max_age_hidden.R
\name{DAISIE_DE_trait_logpEC_max_age_hidden}
\alias{DAISIE_DE_trait_logpEC_max_age_hidden}
\title{Testing function for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpEC_max_age_hidden(
  brts,
  parameter,
  phy,
  traits,
  num_observed_states,
  num_hidden_states,
  sampling_fraction,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  see_ancestral_states = TRUE,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45",
  rhs_func = loglik_hidden_rhs
)
}
\arguments{
\item{brts}{Branching times.}

\item{parameter}{Model parameters.}

\item{phy}{Phylogeny (of class \code{phylo}).}

\item{traits}{Trait states for the species.}

\item{num_observed_states}{Number of observed trait states.}

\item{num_hidden_states}{Number of hidden trait states.}

\item{cond}{Conditioning scheme (default = "proper_cond").}

\item{root_state_weight}{Root state weighting method (default = "proper_weights").}

\item{see_ancestral_states}{Logical; whether to recover ancestral states.}

\item{atol}{Absolute tolerance for ODE integration.}

\item{rtol}{Relative tolerance for ODE integration.}

\item{methode}{Method used for numerical integration (e.g., "ode45").}

\item{missnumspec}{Number of missing species.}

\item{setting_calculation}{Setting used in maximum likelihood estimation.}
}
\description{
This function calculates the likelihood of observing a clade with specified species trait states,
for which the estimated maximum age of colonization is known.
}
\examples{
library(DAISIE)
data("NewZealand_birds_datalist")
datalist <- NewZealand_birds_datalist
i <- 23
phy <- DDD::brts2phylo(datalist[[i]]$branching_times[-c(1, 2)])
traits <- sample(c(0, 1), length(phy$tip.label), replace = TRUE)
sampling_fraction <- sample(c(1, 1), length(phy$tip.label), replace = TRUE)


parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0
)

parameter <- list(2.546591, 2.678781, 0.009326754, 1.008583, matrix(c(0), nrow = 1), 0 )

DAISIE_DE_trait_logpEC_max_age_hidden(
  brts                  = datalist[[i]]$branching_times,
  phy                   = phy,
  traits                = traits,
  sampling_fraction     = sampling_fraction,
  parameter             = parameter,
  num_observed_states   = 2,
  num_hidden_states     = 2,
  cond                  = "proper_cond",
  root_state_weight     = "proper_weights",
  see_ancestral_states  = TRUE,
  atol                  = 1e-10,
  rtol                  = 1e-10,
  methode               = "ode45",
  rhs_func              = loglik_hidden_rhs
)
}
