% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpES_mainland_hidden.R
\name{DAISIE_DE_trait_logpES_mainland_hidden}
\alias{DAISIE_DE_trait_logpES_mainland_hidden}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpES_mainland_hidden(
  brts,
  trait,
  parameter,
  sf = 1,
  num_observed_states,
  num_hidden_states,
  trait_mainland_ancestor = trait_mainland_ancestor,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  see_ancestral_states = TRUE,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45"
)
}
\arguments{
\item{brts}{branching times}

\item{trait}{trait of the species at the tip}

\item{parameter}{parameters}

\item{num_observed_states}{number of observed traits}

\item{num_hidden_states}{number of hidden traits}

\item{trait_mainland_ancestor}{trait of the mainland ancestors}

\item{atol}{absolute tolerance}

\item{rtol}{relative tolerance}

\item{methode}{method of integration}
}
\description{
this function calculates the likelihood of observing a singleton endemic species on an island
with the trait state \code{i}, that coexist on the island with its mainland ancestors.
}
