% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpES_max_min_age_hidden.R
\name{DAISIE_DE_trait_logpES_max_min_age_hidden}
\alias{DAISIE_DE_trait_logpES_max_min_age_hidden}
\title{testing function, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpES_max_min_age_hidden(
  brts,
  trait,
  sf = 1,
  parameter,
  num_observed_states,
  num_hidden_states,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45"
)
}
\arguments{
\item{brts}{branching times}

\item{trait}{trait state of the species at the tip}

\item{parameter}{parameters}

\item{num_observed_states}{number of observed traits}

\item{num_hidden_states}{number of hidden traits}

\item{atol}{absolute tolerance}

\item{rtol}{relative tolerance}

\item{methode}{method of integration}
}
\description{
this function calculates the likelihood of observing a singleton endemic species on an island
with the trait state \code{i}, and for which only the estimated maximum and minimum ages of colonization are known.
}
\examples{
library(DAISIE)
data("Biwa_datalist")
datalist <- Biwa_datalist
sf <- 1
parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0
)

parameter <- list(2.546591, 2.678781, 0.009326754, 1.008583, matrix(c(0), nrow = 1), 0 )


DAISIE_DE_trait_logpES_max_min_age_hidden(
  brts                  = c(4, 3.9999, 0.001),
  trait                 = 0,
  parameter             = parameter,
  num_observed_states   = 2,
  num_hidden_states     = 2,
  atol                  = 1e-10,
  rtol                  = 1e-10,
  methode               = "ode45"
)
}
