% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpNE.R
\name{DAISIE_DE_trait_logpNE}
\alias{DAISIE_DE_trait_logpNE}
\title{This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,}
\usage{
DAISIE_DE_trait_logpNE(
  brts,
  status,
  trait,
  trait_mainland_ancestor = FALSE,
  num_observed_states,
  num_hidden_states,
  parameter,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  use_R = TRUE
)
}
\arguments{
\item{brts}{Branching times.}

\item{trait}{trait state of the species at the tip.}

\item{trait_mainland_ancestor}{Trait state of the species at the stem (mainland ancestor).}

\item{num_observed_states}{Number of observed trait states.}

\item{num_hidden_states}{Number of hidden trait states.}

\item{parameter}{List of model parameters.}

\item{atol}{Absolute tolerance for numerical integration.}

\item{rtol}{Relative tolerance for numerical integration.}

\item{methode}{Numerical integration method (e.g., "ode45").}

\item{missnumspec}{Number of missing species.}
}
\description{
This function calculates the likelihood of observing a non-endemic lineage with specified species trait states,
}
\examples{
library(DAISIE)
library(secsse)
data("Galapagos_datalist")
datalist <- Galapagos_datalist

i <- 2
brts <- datalist[[i]]$branching_times
trait <- 0


parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0
)

parameter <- list(2.546591, 2.678781, 0.009326754, 1.008583, matrix(c(0), nrow = 1), 0 )

DAISIE_DE_trait_logpNE(
  brts                    = brts,
  trait                   = trait,
  status                  = 1,
  parameter               = parameter,
  num_observed_states     = 2,
  num_hidden_states       = 2,
  atol                    = 1e-15,
  rtol                    = 1e-15,
  methode                 = "ode45"
)
}
