% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_n.R
\name{DAISIE_DE_trait_n}
\alias{DAISIE_DE_trait_n}
\title{Convert one rho-sampling lineage likelihood into an n-sampling likelihood}
\usage{
DAISIE_DE_trait_n(log_f, missnumspec, S, step_size = NULL)
}
\description{
Convert one rho-sampling lineage likelihood into an n-sampling likelihood
}
