% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{loglik_choosepar}
\alias{loglik_choosepar}
\title{loglik choosepar temp}
\usage{
loglik_choosepar(
  trparsopt,
  trparsfix,
  idparsopt,
  idparsfix,
  idparslist,
  datalist,
  num_observed_states,
  num_hidden_states,
  cond = cond,
  atol,
  rtol,
  methode,
  rcpp_methode,
  verbose,
  use_Rcpp,
  num_threads
)
}
\arguments{
\item{trparsopt}{transformed parameters to optimize}

\item{trparsfix}{transformed paramaters that remain fixed}

\item{idparsopt}{a numeric vector with the ID of parameters to be estimated.}

\item{idparsfix}{a numeric vector with the ID of the fixed parameters.}

\item{idparslist}{overview of parameters and their values.}

\item{datalist}{a list containing in the first entry: island age and the
number of species not present, then from entry 2 until n it contains
individual datasets for which to estimate the loglikelihood together, each
dataset has the following fields: 1) colonist_name, 2) b ranching_times,
3) stac, 4) missing_species, 5) type1or2, 6) traits, 7) sampling fraction,
8) phylogeny}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{cond}{condition on the existence of a node root: `"maddison_cond"`,
`"proper_cond"` (default). For details, see vignette.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
 `"odeint::runge_kutta_cash_karp54"`, `"odeint::runge_kutta_fehlberg78"`,
 `"odeint::runge_kutta_dopri5"`, `"odeint::bulirsch_stoer"` and
 `"odeint::runge_kutta4"`. Default method is: `"odeint::bulirsch_stoer"`.}

\item{verbose}{sets verbose output; default is `TRUE` when `optimmethod` is
`"simplex"`. If `optimmethod` is set to `"simplex"`, then even if set to
`FALSE`, optimizer output will be shown.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}

\item{num_threads}{number of threads to be used. Default is one thread.}
}
\description{
temporary export for testing
}
