% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_tree.R
\name{loglik_cpp_tree}
\alias{loglik_cpp_tree}
\title{Using hidden traits}
\usage{
loglik_cpp_tree(
  parameter,
  phy,
  traits,
  sampling_fraction,
  num_hidden_states,
  mainland = FALSE,
  trait_mainland_ancestor = NULL,
  atol = 1e-08,
  rtol = 1e-07,
  method = "odeint::bulirsch_stoer",
  use_normalization = TRUE,
  num_threads = 1
)
}
\arguments{
\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{phy}{phylogenetic tree of class \code{phylo}, rooted and with
branch lengths. Alternatively, multiple phylogenetic trees can be provided
as the \code{multiPhylo} class.}

\item{traits}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
\code{vignette("starting_secsse", package = "secsse")}. When providing a
\code{multiPhylo} set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a \code{multiPhylo} object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{mainland}{if TRUE, adjust initial conditions for mainland dynamics}

\item{trait_mainland_ancestor}{if mainland is TRUE, sets the trait of the mainland ancestor}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{method}{integration method used, available are:
\code{"odeint::runge_kutta_cash_karp54"}, \code{"odeint::runge_kutta_fehlberg78"},
\code{"odeint::runge_kutta_dopri5"}, \code{"odeint::bulirsch_stoer"} and
\code{"odeint::runge_kutta4"}. Default method is: \code{"odeint::bulirsch_stoer"}.}

\item{use_normalization}{normalize the density vector during integration,
more accurate but slower (default = TRUE)}

\item{num_threads}{number of threads to be used. Default is one thread.}

\item{rhs_func}{ll function}
}
\description{
Likelihood calculation including hidden traits
}
