% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_hidden_mainland.R
\name{loglik_R_hidden_mainland}
\alias{loglik_R_hidden_mainland}
\title{Using hidden traits}
\usage{
loglik_R_hidden_mainland(
  parameter,
  phy,
  traits,
  sampling_fraction,
  num_hidden_states,
  trait_mainland_ancestor,
  cond = "proper_cond",
  root_state_weight = "proper_weights",
  see_ancestral_states = TRUE,
  atol = 1e-08,
  rtol = 1e-07,
  methode = "ode45",
  use_normalization = TRUE,
  rhs_func = loglik_hidden_rhs_mainland
)
}
\arguments{
\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{phy}{phylogenetic tree of class \code{phylo}, rooted and with
branch lengths. Alternatively, multiple phylogenetic trees can be provided
as the \code{multiPhylo} class.}

\item{traits}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
\code{vignette("starting_secsse", package = "secsse")}. When providing a
\code{multiPhylo} set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a \code{multiPhylo} object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{cond}{condition on the existence of a node root: \code{"maddison_cond"},
\code{"proper_cond"} (default). For details, see vignette.}

\item{root_state_weight}{the method to weigh the states:
\code{"maddison_weights"}, \code{"proper_weights"} (default) or \code{"equal_weights"}.
It can also be specified for the root state: the vector \code{c(1, 0, 0)}
indicates state 1 was the root state. When
using a \code{multiPhylo} object, root_state_weight should be list where each
entry in the list corresponds to the root_state_weight for each tree.}

\item{see_ancestral_states}{Boolean for whether the ancestral states should
be shown? Defaults to \code{FALSE}.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{use_normalization}{normalize the density vector during integration,
more accurate but slower (default = TRUE)}

\item{rhs_func}{ll function}
}
\description{
Likelihood calculation including hidden traits
}
