% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_tree.R
\name{loglik_R_tree}
\alias{loglik_R_tree}
\title{Using hidden traits}
\usage{
loglik_R_tree(
  parameter,
  phy,
  traits,
  sampling_fraction,
  num_hidden_states,
  mainland = FALSE,
  trait_mainland_ancestor = NULL,
  atol = 1e-08,
  rtol = 1e-07,
  methode = "ode45",
  rhs_func = loglik_hidden_rhs
)
}
\arguments{
\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{phy}{phylogenetic tree of class \code{phylo}, rooted and with
branch lengths. Alternatively, multiple phylogenetic trees can be provided
as the \code{multiPhylo} class.}

\item{traits}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
\code{vignette("starting_secsse", package = "secsse")}. When providing a
\code{multiPhylo} set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a \code{multiPhylo} object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{mainland}{if TRUE, adjust initial conditions for mainland dynamics}

\item{trait_mainland_ancestor}{if mainland is TRUE, sets the trait of the mainland ancestor}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{rhs_func}{ll function}
}
\description{
Likelihood calculation including hidden traits
}
