% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_full_lambdas}
\alias{prepare_full_lambdas}
\title{Prepares the entire set of lambda matrices for cla_secsse.
It provides the set of matrices containing all the speciation rates}
\usage{
prepare_full_lambdas(traits, num_concealed_states, lambd_and_modeSpe)
}
\arguments{
\item{traits}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
\code{vignette("starting_secsse", package = "secsse")}. When providing a
\code{multiPhylo} set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{num_concealed_states}{number of concealed states, generally equivalent
to the number of examined states in the dataset.}

\item{lambd_and_modeSpe}{a matrix with the 4 models of speciation possible.}
}
\value{
A list of lambdas, its length would be the same than the number of
trait states * num_concealed_states..
}
\description{
Prepares the entire set of lambda matrices for cla_secsse.
It provides the set of matrices containing all the speciation rates
}
