% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_branch.R
\name{solve_branch}
\alias{solve_branch}
\title{branch solving}
\usage{
solve_branch(
  interval_func,
  initial_conditions,
  time,
  parameter,
  methode,
  atol,
  rtol,
  use_Rcpp = 0
)
}
\arguments{
\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{methode}{used integration method, default is ode45}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree likelihood,
value 2 = use Rcpp for the tree likelihood AND the ODEs along the branches}
}
\description{
solve along branch
}
