% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_branch_cpp.R
\name{solve_branch_cpp}
\alias{solve_branch_cpp}
\title{branch solving}
\usage{
solve_branch_cpp(
  chosen_func,
  initial_conditions,
  time,
  parameter,
  methode,
  atol,
  rtol
)
}
\arguments{
\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{methode}{used integration method, default is ode45}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}
}
\description{
use cpp
}
