% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secsse-package.R
\docType{package}
\name{treeLL-package}
\alias{treeLL}
\alias{treeLL-package}
\title{treeLL: Bare-bones package allowing for implemenation of tree-likelihood}
\description{
Provides a back-bone architecture that allows for integrating a rhs along a tree.
}
\author{
\strong{Maintainer}: Thijs Janzen \email{t.janzen@rug.nl} (\href{https://orcid.org/0000-0002-4162-1140}{ORCID})

Other contributors:
\itemize{
  \item Hanno Hildenbrandt \email{h.hildenbrandt@rug.nl} (\href{https://orcid.org/0000-0002-6784-1037}{ORCID}) [contributor]
}

}
\keyword{internal}
