% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{calc_ml}
\alias{calc_ml}
\title{apply maximum likelihood to a dataset}
\usage{
calc_ml(
  datalist,
  num_observed_states,
  num_hidden_states,
  idparslist,
  idparsopt,
  initparsopt,
  idparsfix,
  parsfix,
  cond = "proper_cond",
  tol = c(1e-04, 1e-05, 1e-07),
  maxiter = 1000 * round((1.25)^length(idparsopt)),
  optimmethod = "simplex",
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  num_cycles = 1,
  verbose = FALSE,
  num_threads = 1,
  atol = 1e-15,
  rtol = 1e-15,
  use_Rcpp = 0
)
}
\arguments{
\item{datalist}{a list containing in the first entry: island age and the
number of species not present, then from entry 2 until n it contains
individual datasets for which to estimate the loglikelihood together, each
dataset has the following fields: 1) colonist_name, 2) b ranching_times,
3) stac, 4) missing_species, 5) type1or2, 6) traits, 7) sampling fraction,
8) phylogeny}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{idparslist}{overview of parameters and their values.}

\item{idparsopt}{a numeric vector with the ID of parameters to be estimated.}

\item{initparsopt}{a numeric vector with the initial guess of the parameters
to be estimated.}

\item{idparsfix}{a numeric vector with the ID of the fixed parameters.}

\item{parsfix}{a numeric vector with the value of the fixed parameters.}

\item{cond}{condition on the existence of a node root: `"maddison_cond"`,
`"proper_cond"` (default). For details, see vignette.}

\item{tol}{A numeric vector with the maximum tolerance of the optimization
algorithm. Default is `c(1e-04, 1e-05, 1e-05)`.}

\item{maxiter}{max number of iterations. Default is
`1000 * round((1.25) ^ length(idparsopt))`.}

\item{optimmethod}{A string with method used for optimization. Default is
`"simplex"`. Alternative is `"subplex"`. Both are called from
[DDD::optimizer()], simplex is implemented natively in the package [DDD],
while subplex is ultimately called from [subplex::subplex()].}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
 `"odeint::runge_kutta_cash_karp54"`, `"odeint::runge_kutta_fehlberg78"`,
 `"odeint::runge_kutta_dopri5"`, `"odeint::bulirsch_stoer"` and
 `"odeint::runge_kutta4"`. Default method is: `"odeint::bulirsch_stoer"`.}

\item{num_cycles}{Number of cycles of the optimization. When set to `Inf`,
the optimization will be repeated until the result is, within the
tolerance, equal to the starting values, with a maximum of 10 cycles.}

\item{verbose}{sets verbose output; default is `TRUE` when `optimmethod` is
`"simplex"`. If `optimmethod` is set to `"simplex"`, then even if set to
`FALSE`, optimizer output will be shown.}

\item{num_threads}{number of threads to be used. Default is one thread.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
ml function
}
