% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_loglik_CS.R
\name{DAISIE_DE_trait_loglik_CS}
\alias{DAISIE_DE_trait_loglik_CS}
\title{apply loglikelihood to a dataset}
\usage{
DAISIE_DE_trait_loglik_CS(
  parameter,
  datalist,
  methode = "lsodes",
  atol = 1e-15,
  rtol = 1e-15,
  num_observed_states,
  num_hidden_states,
  cond = 1,
  num_threads = 1,
  verbose = FALSE,
  use_Rcpp = 0
)
}
\arguments{
\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{methode}{used integration method, default is ode45}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{cond}{condition on the existence of a node root: \code{"maddison_cond"},
\code{"proper_cond"} (default). For details, see vignette.}

\item{num_threads}{number of threads to be used. Default is one thread.}

\item{verbose}{sets verbose output; default is \code{TRUE} when \code{optimmethod} is
\code{"simplex"}. If \code{optimmethod} is set to \code{"simplex"}, then even if set to
\code{FALSE}, optimizer output will be shown.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree likelihood,
value 2 = use Rcpp for the tree likelihood AND the ODEs along the branches}
}
\description{
umbrella function
}
