% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logp0.R
\name{DAISIE_DE_trait_logp0}
\alias{DAISIE_DE_trait_logp0}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logp0(
  datalist,
  parameter,
  atol = 1e-15,
  rtol = 1e-15,
  num_observed_states,
  num_hidden_states,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  use_Rcpp = 0
)
}
\arguments{
\item{datalist}{a list containing in the first entry: island age and the
number of species not present, then from entry 2 until n it contains
individual datasets for which to estimate the loglikelihood together, each
dataset has the following fields: 1) colonist_name, 2) b ranching_times,
3) stac, 4) missing_species, 5) type1or2, 6) traits, 7) sampling fraction,
8) phylogeny}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
 `"odeint::runge_kutta_cash_karp54"`, `"odeint::runge_kutta_fehlberg78"`,
 `"odeint::runge_kutta_dopri5"`, `"odeint::bulirsch_stoer"` and
 `"odeint::runge_kutta4"`. Default method is: `"odeint::bulirsch_stoer"`.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
This function compute the likelihood that all species that colonize the island
have gone extinct prior to the present.
}
\examples{
#Load DAISIE package and data
library(DAISIE)
data("Galapagos_datalist")
datalist <- Galapagos_datalist

parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0
)

# Compute log‐likelihood under the DE‐trait model
DAISIE_DE_trait_logp0(
  datalist            = datalist,
  parameter           = parameter,
  num_observed_states     = 2,
  num_hidden_states       = 2,
  atol                = 1e-5,
  rtol                = 1e-5,
  methode             = "ode45"
)

parameter <- list(
  c(2.546591, 2.546591, 2.546591, 2.546591),
  c(2.678781, 2.678781, 2.678781, 2.678781),
  c(0.009326754, 0.009326754, 0.009326754, 0.009326754),
  c(1.008583, 1.008583, 1.008583, 1.008583),
  matrix(c(
    rep(0, 16)
  ), nrow = 4),
  0
)
# Compute log‐likelihood under the DE‐trait model
DAISIE_DE_trait_logp0(
  datalist            = datalist,
  parameter           = parameter,
  num_observed_states     = 2,
  num_hidden_states       = 2,
  atol                = 1e-15,
  rtol                = 1e-15,
  methode             = "ode45"
)
}
