% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpES.R
\name{DAISIE_DE_trait_logpES}
\alias{DAISIE_DE_trait_logpES}
\title{Testing function for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpES(
  brts,
  status,
  trait,
  sampling_fraction,
  num_observed_states,
  num_hidden_states,
  trait_mainland_ancestor = NA,
  parameter,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  use_Rcpp = 0
)
}
\arguments{
\item{brts}{branching times, being: island age, colonization time, branching
times}

\item{status}{indicator for type of inference required}

\item{trait}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
`vignette("starting_secsse", package = "secsse")`. When providing a
`multiPhylo` set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a `multiPhylo` object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{trait_mainland_ancestor}{if mainland is TRUE, sets the trait of the
mainland ancestor}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
 `"odeint::runge_kutta_cash_karp54"`, `"odeint::runge_kutta_fehlberg78"`,
 `"odeint::runge_kutta_dopri5"`, `"odeint::bulirsch_stoer"` and
 `"odeint::runge_kutta4"`. Default method is: `"odeint::bulirsch_stoer"`.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
This function calculates the likelihood of observing a clade with specified
species trait states, given known colonization time. It is designed for
comparison with DAISIE-based models.
}
\examples{
library(DAISIE)
library(secsse)
data("Galapagos_datalist")
datalist <- Galapagos_datalist

i <- 7
brts <- datalist[[i]]$branching_times
trait <- 0
sampling_fraction <- c(1,1)


parameter <- list(
  c(2.546591, 2.546591),
  c(2.678781, 2.678781),
  c(0.009326754, 0.009326754),
  c(1.008583, 1.008583),
  matrix(c(
    rep(0, 4)
  ), nrow = 2),
  0
)

DAISIE_DE_trait_logpES(
  brts                    = brts,
  trait                   = 0,
  status                  = 2,
  sampling_fraction       = sampling_fraction,
  parameter               = parameter,
  trait_mainland_ancestor = c(1, 0),
  num_observed_states     = 2,
  num_hidden_states       = 1,
  atol                    = 1e-15,
  rtol                    = 1e-15,
  methode                 = "ode45"
)
}
