% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpES_max_min_age_hidden.R
\name{DAISIE_DE_trait_logpES_max_min_age_hidden}
\alias{DAISIE_DE_trait_logpES_max_min_age_hidden}
\title{testing function, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpES_max_min_age_hidden(
  brts,
  trait,
  status,
  sampling_fraction = 1,
  parameter,
  num_observed_states,
  num_hidden_states,
  atol = 1e-10,
  rtol = 1e-10,
  methode = "ode45",
  use_Rcpp = 0
)
}
\arguments{
\item{brts}{branching times, being: island age, colonization time, branching times}

\item{trait}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
\code{vignette("starting_secsse", package = "secsse")}. When providing a
\code{multiPhylo} set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{status}{indicator for type of inference required}

\item{sampling_fraction}{vector that states the sampling proportion per
trait state. It must have as many elements as there are trait states. When
using a \code{multiPhylo} object, sampling fraction should be list where each
entry in the list corresponds to the sampling proportion for each tree.}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree likelihood,
value 2 = use Rcpp for the tree likelihood AND the ODEs along the branches}
}
\description{
this function calculates the likelihood of observing a singleton endemic species on an island
with the trait state \code{i}, and for which only the estimated maximum and minimum ages of colonization are known.
}
\examples{
library(DAISIE)
data("Biwa_datalist")
datalist <- Biwa_datalist
sf <- 1
parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0
)

parameter <- list(2.546591, 2.678781, 0.009326754, 1.008583, matrix(c(0), nrow = 1), 0 )


DAISIE_DE_trait_logpES_max_min_age_hidden(
  brts                  = c(4, 3.9999, 0.001),
  trait                 = 0,
  status                = 9,
  parameter             = parameter,
  num_observed_states   = 1,
  num_hidden_states     = 1,
  atol                  = 1e-10,
  rtol                  = 1e-10,
  methode               = "ode45"
)
}
