% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_trait_logpNE_max_min_age_hidden.R
\name{DAISIE_DE_trait_logpNE_max_min_age_hidden}
\alias{DAISIE_DE_trait_logpNE_max_min_age_hidden}
\title{testing fuction, for comparison with DAISIE}
\usage{
DAISIE_DE_trait_logpNE_max_min_age_hidden(
  brts,
  trait,
  status,
  parameter,
  trait_mainland_ancestor = NA,
  num_observed_states,
  num_hidden_states,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  use_Rcpp = 0
)
}
\arguments{
\item{brts}{branching times, being: island age, colonization time, branching
times}

\item{trait}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips. For help, see
\code{vignette("starting_secsse", package = "secsse")}. When providing a
\code{multiPhylo} set of multiple phylognies, traits should be a list where
each entry in the list corresponds to the matching phylogeny on that
position.}

\item{status}{indicator for type of inference required}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{trait_mainland_ancestor}{if mainland is TRUE, sets the trait of the
mainland ancestor}

\item{num_observed_states}{number of observed states}

\item{num_hidden_states}{number of hidden states}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
\code{"odeint::runge_kutta_cash_karp54"}, \code{"odeint::runge_kutta_fehlberg78"},
\code{"odeint::runge_kutta_dopri5"}, \code{"odeint::bulirsch_stoer"} and
\code{"odeint::runge_kutta4"}. Default method is: \code{"odeint::bulirsch_stoer"}.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
this function calculates the likelihood of observing a singleton species on
an island with the trait state \code{i}, either non-endemic or rendered endemic
by a trait change, and for which only the estimated maximum age of
colonization is known.
}
\examples{
# load DAISIE package and data
library(DAISIE)

parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    1,    0.5,  0,
    0,    0,    0.002,0.005,
    rep(0, 8)
  ), nrow = 4),
  0, c(0,1)
)
status <- 8
parameter <- list(2.546591, 2.678781, 0.009326754, 1.008583,
                  matrix(c(0), nrow = 1), 0 )


DAISIE_DE_trait_logpNE_max_min_age_hidden(
  brts                  = c(4, 3.999, 0.0001),
  trait                 = 0,
  status                = 8,
  parameter             = parameter,
  num_observed_states   = 2,
  num_hidden_states     = 2,
  atol                  = 1e-15,
  rtol                  = 1e-15,
  methode               = "ode45"
)
}
