% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{loglik_choosepar}
\alias{loglik_choosepar}
\title{loglik choosepar temp}
\usage{
loglik_choosepar(
  trparsopt,
  trparsfix,
  idparsopt,
  idparsfix,
  idparslist,
  datalist,
  num_observed_states,
  num_hidden_states,
  cond = cond,
  atol,
  rtol,
  methode,
  verbose,
  use_R
)
}
\description{
temporary export for testing
}
