% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_branch.R
\name{solve_branch}
\alias{solve_branch}
\title{branch solving}
\usage{
solve_branch(
  interval_func,
  initial_conditions,
  time,
  parameter,
  trait_mainland_ancestor,
  methode = "ode45",
  rcpp_methode = "odeint::bulirsch_stoer",
  atol,
  rtol,
  use_Rcpp = 0
)
}
\arguments{
\item{interval_func}{chosen function for interval, can also be string if
using Rcpp}

\item{initial_conditions}{vector of initial conditions}

\item{time}{vector with two time points}

\item{parameter}{list where first vector represents lambdas, the second
mus and the third transition rates.}

\item{trait_mainland_ancestor}{if mainland is TRUE, sets the trait of the
mainland ancestor}

\item{methode}{used integration method, default is ode45}

\item{rcpp_methode}{Used integration methode when using Rcpp to integrate,
options available are:
 `"odeint::runge_kutta_cash_karp54"`, `"odeint::runge_kutta_fehlberg78"`,
 `"odeint::runge_kutta_dopri5"`, `"odeint::bulirsch_stoer"` and
 `"odeint::runge_kutta4"`. Default method is: `"odeint::bulirsch_stoer"`.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{use_Rcpp}{default is 0: use no Rcpp. Value 1 = use Rcpp for the tree
likelihood, value 2 = use Rcpp for the tree likelihood AND the ODEs along
the branches}
}
\description{
solve along branch
}
